Improve Roku channel launch support and control server/lambda custom variables.
This will update lambda server and control server to allow for
channel names to be sent dynamically from the skill and looked up
in a channel list.

The APP_ID will be moved to the serverinfo_example.js and a new
serverinfo_example.js is created for the control server
adding in the rokuaddress, server listening port, and
authorization password.

.gitignore has an ignore added for the RokuControlServer/serverinfo.js
and any generated .zip files in the RokuLamda dir.